refactor(standards): remove the AuthSingleSigAcl auth component
Its exempt-procedure-list mechanism let fee-charging accounts be
drained: anyone could call an exempt procedure without a signature,
but the account still paid the transaction fee. The team decided to
remove the component rather than fix it now (see issue discussion).

The plain AuthSingleSig component (every call requires a signature)
remains available and now backs the "singlesig user faucet" factories.

Closes #3360
