--- crates/miden-standards/asm/components/auth/singlesig_acl/miden-project.toml
@@ -1,13 +1,0 @@
-[package]
-name              = "miden-standards-auth-singlesig-acl"
-version.workspace = true
-
-[lib]
-kind      = "account-component"
-namespace = "miden::standards::components::auth::singlesig_acl"
-path      = "singlesig_acl.masm"
-
-[dependencies]
-miden-core.workspace      = true
-miden-protocol.workspace  = true
-miden-standards.workspace = true
