+++ crates/miden-standards/asm/components/miden-project.toml
@@ -11,7 +11,6 @@ members = [
   "auth/network_account",
   "auth/no_auth",
   "auth/singlesig",
-  "auth/singlesig_acl",
   "faucets/fungible_faucet",
   "faucets/non_fungible_faucet",
   "faucets/policies/burn/allow_all",
