+++ crates/miden-standards/src/account/auth/approver.rs
@@ -13,7 +13,7 @@ use miden_protocol::errors::AccountError;
 ///
 /// Note: an approver using [`AuthScheme::EcdsaK256Keccak`] discloses its public key and signature
 /// at proving time and therefore does not provide public-key privacy, regardless of the component
-/// it is used in (single-sig, ACL, multisig, or guarded multisig). See
+/// it is used in (single-sig, multisig, or guarded multisig). See
 /// [`AuthScheme::EcdsaK256Keccak`] for details, and prefer [`AuthScheme::Falcon512Poseidon2`] if
 /// signer-key privacy is required.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
