+++ crates/miden-standards/src/account/auth/mod.rs
@@ -10,9 +10,6 @@ pub use no_auth::NoAuth;
 mod singlesig;
 pub use singlesig::AuthSingleSig;
 
-mod singlesig_acl;
-pub use singlesig_acl::{AuthSingleSigAcl, AuthSingleSigAclConfig};
-
 mod multisig;
 pub use multisig::{AuthMultisig, AuthMultisigConfig};
