+++ crates/miden-testing/tests/auth/mod.rs
@@ -1,7 +1,6 @@
 mod fee_payment;
 
 mod singlesig;
-mod singlesig_acl;
 
 mod multisig;
