+++ crates/miden-testing/tests/auth/singlesig.rs
@@ -1,19 +1,35 @@
 use assert_matches::assert_matches;
 use miden_processor::ExecutionError;
-use miden_protocol::Word;
+use miden_processor::crypto::random::RandomCoin;
 use miden_protocol::account::auth::{AuthScheme, AuthSecretKey};
 use miden_protocol::account::{
     Account,
     AccountBuilder,
     AccountComponent,
+    AccountId,
     AccountStorage,
     AccountType,
 };
+use miden_protocol::asset::{AssetAmount, FungibleAsset, TokenSymbol};
 use miden_protocol::errors::MasmError;
 use miden_protocol::note::Note;
 use miden_protocol::transaction::RawOutputNote;
-use miden_standards::account::auth::AuthSingleSig;
+use miden_protocol::{Felt, Word};
+use miden_standards::account::auth::{Approver, AuthSingleSig};
+use miden_standards::account::faucets::{
+    Description,
+    FungibleFaucet,
+    TokenName,
+    create_singlesig_user_fungible_faucet,
+};
+use miden_standards::account::policies::{
+    BurnPolicy,
+    MintPolicy,
+    TokenPolicyManager,
+    TransferPolicy,
+};
 use miden_standards::code_builder::CodeBuilder;
+use miden_standards::note::BurnNote;
 use miden_standards::testing::account_component::MockAccountComponent;
 use miden_standards::testing::note::NoteBuilder;
 use miden_testing::{Auth, MockChain, assert_transaction_executor_error};
@@ -181,6 +197,132 @@ async fn test_singlesig_auth_rejects_rotated_key_signature(
         },
         _ => unreachable!("only the two rstest cases are parameterized"),
     }
+
+    Ok(())
+}
+
+// BURN NOTE AGAINST SINGLESIG USER FAUCET
+// ================================================================================================
+
+/// Sets up a singlesig user fungible faucet (built via the production
+/// `create_singlesig_user_fungible_faucet` factory) with an existing token supply, and a BURN
+/// note targeting it. Returns (faucet account, mock chain, burn note, authenticator).
+fn setup_singlesig_user_faucet_with_burn_note(
+    auth_scheme: AuthScheme,
+) -> anyhow::Result<(Account, MockChain, Note, BasicAuthenticator)> {
+    let mut rng = ChaCha20Rng::from_seed(Default::default());
+    let sec_key = AuthSecretKey::with_scheme_and_rng(auth_scheme, &mut rng)?;
+    let pub_key = sec_key.public_key().to_commitment();
+    let authenticator = BasicAuthenticator::new(&[sec_key]);
+
+    let auth_component = AuthSingleSig::new(Approver::new(pub_key, auth_scheme));
+
+    let policy_manager = TokenPolicyManager::builder()
+        .active_mint_policy(MintPolicy::allow_all())
+        .active_burn_policy(BurnPolicy::allow_all())
+        .active_send_policy(TransferPolicy::allow_all())
+        .active_receive_policy(TransferPolicy::allow_all())
+        .build();
+
+    let faucet = FungibleFaucet::builder()
+        .name(TokenName::new("polygon")?)
+        .symbol(TokenSymbol::try_from("POL")?)
+        .decimals(2)
+        .max_supply(AssetAmount::new(1000)?)
+        .token_supply(AssetAmount::new(100)?)
+        .description(Description::new("A polygon token")?)
+        .build()?;
+
+    let faucet_account = create_singlesig_user_fungible_faucet(
+        [42u8; 32],
+        faucet,
+        auth_component,
+        policy_manager,
+        AccountType::Public,
+    )?;
+    // The factory builds a new account (nonce 0, carrying a seed); MockChain genesis accounts
+    // must be existing (nonce != 0, no seed), so re-wrap the built code/storage/vault as such.
+    let faucet_account = Account::new(
+        faucet_account.id(),
+        faucet_account.vault().clone(),
+        faucet_account.storage().clone(),
+        faucet_account.code().clone(),
+        Felt::ONE,
+        None,
+    )?;
+
+    let sender = AccountId::builder().account_type(AccountType::Private).build_with_seed([3; 32]);
+    let asset = FungibleAsset::new(faucet_account.id(), 10)?;
+    let mut note_rng = RandomCoin::new([Felt::from(7u32); 4].into());
+    let burn_note: Note = BurnNote::builder()
+        .sender(sender)
+        .asset(asset)
+        .generate_serial_number(&mut note_rng)
+        .build()?
+        .into();
+
+    let mut builder = MockChain::builder();
+    builder.add_account(faucet_account.clone())?;
+    builder.add_output_note(RawOutputNote::Full(burn_note.clone()));
+    let mock_chain = builder.build()?;
+
+    Ok((faucet_account, mock_chain, burn_note, authenticator))
+}
+
+/// A BURN note targeted at a plain `AuthSingleSig` user faucet cannot be consumed without a
+/// signature: unlike the removed `AuthSingleSigAcl`, `AuthSingleSig` has no exempt list, so every
+/// call - including one that only reaches `receive_and_burn` - requires authentication.
+#[rstest]
+#[case::ecdsa(AuthScheme::EcdsaK256Keccak)]
+#[case::falcon(AuthScheme::Falcon512Poseidon2)]
+#[tokio::test]
+async fn test_singlesig_burn_note_against_user_faucet_requires_signature(
+    #[case] auth_scheme: AuthScheme,
+) -> anyhow::Result<()> {
+    let (faucet_account, mock_chain, burn_note, _authenticator) =
+        setup_singlesig_user_faucet_with_burn_note(auth_scheme)?;
+
+    let result = mock_chain
+        .build_transaction(faucet_account.id())
+        .authenticated_input_note(burn_note.id())
+        .authenticator(None)
+        .build()?
+        .execute()
+        .await;
+
+    assert_matches!(result, Err(TransactionExecutorError::MissingAuthenticator));
+
+    Ok(())
+}
+
+/// The same BURN note succeeds once the faucet owner's signature is supplied, confirming that a
+/// faucet built by `create_singlesig_user_fungible_faucet` can still burn assets, just not for
+/// free. `AuthSingleSig` bumps the nonce unconditionally regardless of whether state changed
+/// (see `singlesig.masm`), so a nonce-only assertion would not prove the burn ran; instead this
+/// asserts the faucet's token supply actually dropped by the burned amount.
+#[rstest]
+#[case::ecdsa(AuthScheme::EcdsaK256Keccak)]
+#[case::falcon(AuthScheme::Falcon512Poseidon2)]
+#[tokio::test]
+async fn test_singlesig_burn_note_against_user_faucet_succeeds_with_signature(
+    #[case] auth_scheme: AuthScheme,
+) -> anyhow::Result<()> {
+    let (faucet_account, mock_chain, burn_note, authenticator) =
+        setup_singlesig_user_faucet_with_burn_note(auth_scheme)?;
+
+    let executed_transaction = mock_chain
+        .build_transaction(faucet_account.id())
+        .authenticated_input_note(burn_note.id())
+        .authenticator(Some(authenticator))
+        .build()?
+        .execute()
+        .await?;
+
+    let mut updated_account = faucet_account.clone();
+    updated_account.apply_patch(executed_transaction.account_patch())?;
+    let updated_faucet = FungibleFaucet::try_from(updated_account.storage())?;
+
+    assert_eq!(updated_faucet.token_supply(), AssetAmount::new(90)?);
 
     Ok(())
 }
